+++ adapter/__init__.py
@@ -4,19 +4,21 @@ Adapter module for image generation plugin
 """
 
 from .agnes_ai_adapter import AgnesAIAdapter
+from .custom_http_adapter import CustomHTTPAdapter
 from .gemini_adapter import GeminiAdapter
-from .gemini_openai_adapter import GeminiOpenAIAdapter
 from .gitee_ai_adapter import GiteeAIAdapter
 from .grok_adapter import GrokAdapter
 from .jimeng2api_adapter import Jimeng2APIAdapter
+from .openai_chat_adapter import OpenAIChatAdapter
 from .openai_adapter import OpenAIAdapter
 from .siliconflow_adapter import SiliconFlowAdapter
 from .volcengine_ark_adapter import VolcengineArkAdapter
 
 __all__ = [
     "AgnesAIAdapter",
+    "CustomHTTPAdapter",
     "GeminiAdapter",
-    "GeminiOpenAIAdapter",
+    "OpenAIChatAdapter",
     "GiteeAIAdapter",
     "OpenAIAdapter",
     "SiliconFlowAdapter",